Consolidate stat var (group) related API
