+++ proto/v1/variable.proto
@@ -16,6 +16,26 @@ syntax = "proto3";
 option go_package = "./proto";
 package datacommons.v1;
 
+
+message VariablesRequest {
+  string entity = 1;
+}
+
+message VariablesResponse {
+  string entity = 1;
+  repeated string variables = 2;
+}
+
+message BulkVariablesRequest {
+  repeated string entities = 1;
+  // If set, return the union of all variables of all entities.
+  bool union = 2;
+}
+
+message BulkVariablesResponse {
+  repeated VariablesResponse data = 1;
+}
+
 message VariableAncestorsRequest {
   // DCID of a stat var (group)
   string node = 1;
