--- proto/v1/variable_group.proto
@@ -1,30 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-option go_package = "./proto";
-package datacommons.v1;
-
-import "stat_var.proto";
-
-message VariableGroupsRequest {
-  // (optional) When specified, only return the variable groups that the entities
-  // have data for.
-  repeated string constrained_entities = 1;
-}
-
-message VariableGroupsResponse {
-  // Key is StatVar Group ID.
-  map<string, StatVarGroupNode> data = 1;
-}
