--- proto/v1/variables.proto
@@ -1,36 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-option go_package = "./proto";
-package datacommons.v1;
-
-message VariablesRequest {
-  string entity = 1;
-}
-
-message VariablesResponse {
-  string entity = 1;
-  repeated string variables = 2;
-}
-
-message BulkVariablesRequest {
-  repeated string entities = 1;
-  // If set, return the union of all variables of all entities.
-  bool union = 2;
-}
-
-message BulkVariablesResponse {
-  repeated VariablesResponse data = 1;
-}
